update src and eslintrc for skip errored lint checks
